+++ scripts/generate-component.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+node ./scripts/generateComponent.js "$@"
+yarn generate-tests