+++ scripts/generateComponent.js
@@ -3,7 +3,9 @@ const fs = require('fs');
 const { logInfo, logError, isComponentDirectory } = require('./scriptUtils');
 
 const COMPONENTS_PATH = './src/components';
-const currentComponents = fs.readdirSync(COMPONENTS_PATH).map(name => name)
+const currentComponents = fs
+  .readdirSync(COMPONENTS_PATH)
+  .map(name => name)
   .filter(isComponentDirectory(COMPONENTS_PATH));
 
 const componentName = process.argv[process.argv.indexOf('--name') + 1];
@@ -19,15 +21,27 @@ const ComponentBoilerplate = `import React from 'react';
 import PropTypes from 'prop-types';
 ${stylesBoilterplate}
 
-const ${componentName} = (props) => (
+const Wrapper = styled.span\`\`;
 
-);
+const ${componentName} = ({ name }) => (<Wrapper>Hello, I am {name}</Wrapper>);
 
-${componentName}.defaultProps = {};
+${componentName}.defaultProps = {
+  name: '${componentName}',
+};
 
-${componentName}.propTypes = {};
+${componentName}.propTypes = {
+  // Please remove unused prop types
+  name: PropTypes.string,
+};
 
 export default ${componentName};
+`;
+
+const ComponentDocumentationBoilerplate = `
+${componentName} component:
+\`\`\`jsx
+<${componentName} /> 
+\`\`\`
 `;
 
 // Create new folder.
@@ -39,9 +53,12 @@ fs.writeFileSync(`${componentPath}/${componentName}.js`, ComponentBoilerplate);
 fs.writeFileSync(`${componentPath}/index.js`, `export default from './${componentName}';\n`);
 
 // Create markdown file.
-fs.writeFileSync(`${componentPath}/${componentName}.md`, '');
+fs.writeFileSync(`${componentPath}/${componentName}.md`, ComponentDocumentationBoilerplate);
 
 // Add export in src/components/index.js
-fs.appendFileSync('./src/index.js', `export ${componentName} from './components/${componentName}';\n`);
+fs.appendFileSync(
+  './src/index.js',
+  `export ${componentName} from './components/${componentName}';\n`
+);
 
 logInfo(`Your component can be found in ${componentPath}`);